+++ ui/src/pages/landingPage/LandingPage.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react';
 import { FC } from 'react';
 import styled from 'styled-components';
 import { Colors } from '../../styledHelpers/Colors';
-import { CreateAccountButton } from './common';
-import RegistrationForm from './RegistrationForm';
+import { CreateAccountButton } from '../components/common';
+import RegistrationForm from '../components/registrationForm/RegistrationForm';
 
 const Wrapper = styled.div`
     background-color: rgba(0,0,0,.4);
